Add Constant For Nil UUID
As defined by RFC 4122, this UUID is the UUID with all bits set to 0. Pragmatically, it is a useful UUID to use in instances where one requires a known UUID value, but would prefer not to generate one. This use case often arises in testing, but it is not limited to that scope.